% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mask.R
\name{df_mask}
\alias{df_mask}
\title{Mask person-sensitive counts}
\usage{
df_mask(x)
}
\arguments{
\item{x}{data.table.}
}
\value{
data.table
}
\description{
Masks person-sensitive counts in a 'tblsummary', or more precisely, adds
a variable to the input table, that can be used to do the masking.
}
\examples{
\dontrun{
  x <- data.frame(
    var = c("A", "B", "B", rep("C", 5L))
  )
  tbl1 <- summarize_df(x, var = "var")
  tbl2 <- df_mask(tbl1)
}
}
\keyword{internal}
