% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_binary}
\alias{.is_binary}
\title{Checks if a vector is binary(-ish)}
\usage{
.is_binary(x, tol = .Machine$double.eps^0.5, allow_na = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector}

\item{tol}{Numeric value. Tolerance.}

\item{allow_na}{A logical. Should NA values in \code{x} be accepted as binary?}
}
\value{
A logical
}
\description{
Check if each element of a vector can be interpreted as binary.
}
\details{
A vector element is deemed to be binary if it is of type "logical", or
if it is a numeric value that is indistinguishable from zero or one based
on a tolerence given in \code{tol}.
}
\examples{
\dontrun{
  .is_binary(c(1,0,1))
}
}
\keyword{internal}
