% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{.is_var_name}
\alias{.is_var_name}
\title{Check if variable exists in data.frame.}
\usage{
.is_var_name(x, var)
}
\arguments{
\item{var}{String. Variable name.}

\item{x.}{Input data.frame.}
}
\value{
invisible(NULL) or an error if an input argument is invalid.
}
\description{
Checks that a given variable name is a variable name in input data.frame.
}
\examples{
\dontrun{
  x <- data.frame(var1 = 1:2)
  .is_var_name(x, "var1")
}
}
\keyword{internal}
