% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tbl.R
\name{.mask_vector_element}
\alias{.mask_vector_element}
\title{TODO}
\usage{
.mask_vector_element(x, x_sum, min_mask = 1L, max_mask = 4L, avg_mask = 1L)
}
\arguments{
\item{x}{a}

\item{x_sum}{a}

\item{min_mask}{a}

\item{max_mask}{a}

\item{avg_mask}{a}
}
\value{
a
}
\description{
TODO
}
\examples{
\dontrun{TODO}
}
\keyword{internal}
