% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{.summarize_var}
\alias{.summarize_var}
\title{Summarizes a variable}
\usage{
.summarize_var(
  x,
  var,
  type = NULL,
  weight = NULL,
  by = NULL,
  by_total_val = ".all"
)
}
\arguments{
\item{x}{Input data.frame.}

\item{var}{String. Name of variable in \code{x} to summarize.}

\item{type}{String. Variable type of \code{var}: "bin" (binary),
"cont" (continuous), "cat" (categorical). Default is NULL, meaning that
the type is automatically deduced from the data. See details for more
information.}

\item{weight}{String. Name of variable in \code{x} that contains
observation weights. Must be a numeric variable, with non-negative
numbers. By default (\code{weight} = NULL), all weights are set to one.}

\item{by}{String. Name of variable to group by while summarizing. Default is
NULL, ie no grouping variable is used.}

\item{by_total_val}{String. Name given to overall by-group.}
}
\value{
A data.table containing a descriptive summary of \code{var}.
}
\description{
Creates a descriptive summary of a variable in a data.frame.
}
\details{
If \code{type} is NULL, the variable type of \code{var} will be deduced from the data:
\enumerate{
\item If the variable takes "binary" values, eg is a logical vector, or a numeric
vector with zero and one's, the \code{type} is set to "bin".
\item Else, if the variable is numeric then \code{type} is set to "cont".
\item Else, \code{type} is set to "cat".
}
}
\examples{
\dontrun{
  x <- data.frame(var = 1:10)
  .summarize_var(x, "var")
}
}
\keyword{internal}
