% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mask.R
\name{.update_mask_flags}
\alias{.update_mask_flags}
\title{title}
\usage{
.update_mask_flags(x, x_sum, min_mask = 1L, max_mask = 4L, avg_mask = 1L)
}
\arguments{
\item{x}{a}

\item{x_sum}{a}

\item{min_mask}{a}

\item{max_mask}{a}

\item{avg_mask}{a}
}
\value{
a
}
\description{
TODO:
\itemize{
\item Loop until nothing more is flagged for masking.
\item Change function so that input is a vector and a corresponding
vector with flags for which elements are do be flagged (so far).
the function then updates the flag vector and returns that updated
vector of flags
\item Function should also be able to take a vector with flags for values that
are protected from masking, eg NA's (probably not relevant in this package)
or zeros that are known to be zeros and therefore are not valid to mask
(very relevant, something we want to be able to control)
\item x_sum is going to be obsolete, we can simply sum x since it is not going to
be a partially masked vector anymore? Maybe remove any NA's just to make it
compatible with data with true NA's?
}
}
\examples{
\dontrun{TODO}
}
\keyword{internal}
