% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_tblsummary}
\alias{.validate_tblsummary}
\title{Validates a 'tblsummary' object}
\usage{
.validate_tblsummary(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
invisible(\code{x}).
}
\description{
Validate a 'tblsummary' object. Invisibly returns \code{x} or throws an error.
}
\examples{
\dontrun{
  x <- data.frame(var = 1:5)
  tbl <- .summarize_var(x, var = "var")
  .validate.tblsummary(tbl)
  .validate_tblsummary(data.frame())
}
}
\keyword{:}
\keyword{internal}
