% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.w_inputcheck}
\alias{.w_inputcheck}
\title{Check input for internal functions}
\usage{
.w_inputcheck(x, w, na.rm)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers or NULL}

\item{na.rm}{Boolean.}
}
\description{
Utility function for checking input to internal functions of weighted
versions of statistical functions
}
\examples{
\dontrun{
 .w_inputcheck(1:5, NULL, TRUE)
}
}
\keyword{:}
\keyword{internal}
