% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.winputcheck}
\alias{.winputcheck}
\title{Check common arguments}
\usage{
.winputcheck(x, w, na.rm)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}. Default is NULL, ie no weights are used.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
invisible(NULL) or an error if an argument is invalid
}
\description{
Check that input to common arguments of internal weighted statistical
summary functions are valid.
}
\examples{
\dontrun{
 .winputcheck(1:5, NULL, TRUE)
}
}
\keyword{:}
\keyword{internal}
