% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wmean}
\alias{.wmean}
\title{Weighted arithmetic mean}
\usage{
.wmean(x, w = rep(1L, length(x)), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
A numeric value.
}
\description{
Calculates the weighted arithmetic mean of a numeric vector \code{x} with
associated weights \code{w}.
}
\details{
Formula used:
\deqn{\bar{x} = \frac{\sum_{i = 1}^n{w_i*x_i}}{\sum_{i = 1}^n{w_i}}}
}
\examples{
\dontrun{
  .wmean(x = 1:2, w = 1:2)
}
}
\keyword{internal}
