% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wquantile}
\alias{.wquantile}
\title{Weighted quantile}
\usage{
.wquantile(x, w, prob, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with \code{x}.}

\item{prob}{Numeric value. Quantile to estimate.}

\item{na.rm}{Boolean. Should NA values in \code{x} and values in \code{x} with
NA weights in \code{w} be removed?}
}
\value{
A numeric.
}
\description{
Calculates weighted type 7 quantile. Simplified version of code from
https://arxiv.org/pdf/2304.07265.pdf.
TODO: need to read up on this to make sure everything makes sense.
}
\examples{
\dontrun{TODO}
}
\keyword{:}
\keyword{internal}
