% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wsum}
\alias{.wsum}
\title{Weighted sum of vector elements}
\usage{
.wsum(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with \code{x}.}

\item{na.rm}{Boolean. Should NA values in \code{x} and values in \code{x} with
NA weights in \code{w} be removed?}
}
\value{
A numeric.
}
\description{
Calculates the weighted sum of a numeric vector \code{x} with associated
weights \code{w}.
}
\examples{
\dontrun{
   x <- 1:5
   w <- rep(2, 5)
   .wsum(x, w)
}
}
\keyword{internal}
