% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tbl.R
\name{mask_tbl}
\alias{mask_tbl}
\title{Mask person-sensitive counts}
\usage{
mask_tbl(x)
}
\arguments{
\item{x}{data.table.}
}
\value{
data.table
}
\description{
Masks person-sensitive counts in a 'tblsummary', or more precisely, adds
a variable to the input table, that can be used to do the masking.
}
\examples{
x <- data.frame(
  var = c("A", "B", "B", rep("C", 5L))
)
tbl1 <- summarize_df(x, var = "var")
tbl2 <- mask_tbl(tbl1)
}
