% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{summarize_df}
\alias{summarize_df}
\title{Summarize variables in a data.frame}
\usage{
summarize_df(
  x,
  vars,
  types = NULL,
  weight = NULL,
  by = NULL,
  by_total_val = ".all"
)
}
\arguments{
\item{x}{Input data.frame.}

\item{vars}{Character vector. Names of variables in \code{x} to summarize.}

\item{types}{Optional named character vector. Variable types of variabls in
\code{vars}: "bin" (binary), "cont" (continuous), "cat" (categorical). Default
is NULL, meaning that the types of \code{vars} is automatically deduced from
the data. Can be specified for only some variables, implicitly setting the
type for the remaining variables to NULL.}

\item{weight}{String. Optional name of variable in \code{x} that contains
observation weights. The weights must be non-negative numbers.}

\item{by}{String. Optional name of variable to summarize by.}

\item{by_total_val}{String. Name given to overall by-group.}
}
\value{
A data.table containing a descriptive summary of \code{vars}.
}
\description{
Returns a descriptive summary of variables in a data.frame.
}
\examples{
x <- data.frame(var1 = 1:10, var2 = c(rep("A", 5), rep("B", 5)))
summarize_df(
  x,
  vars = c("var1", "var2"),
  types = c(var1 = "cont", var2 = "cat")
)
}
