% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{.summarize_var}
\alias{.summarize_var}
\title{Summarize a variable}
\usage{
.summarize_var(
  x,
  var,
  type = NULL,
  weight = NULL,
  by = NULL,
  by_total_val = ".all"
)
}
\arguments{
\item{x}{Input data.frame.}

\item{var}{String. Name of variable in \code{x} to summarize.}

\item{type}{String. Variable type of \code{var}: "bin" (binary),
"cont" (continuous), "cat" (categorical). Default is NULL, meaning that
the type is automatically deduced from the data. See details for more
information.}

\item{weight}{String. Name of variable in \code{x} that contains
observation weights. Must be a numeric variable, with non-negative
real values. By default (\code{weight} = NULL), all weights are set to one.}

\item{by}{String. Name of variable to group by while summarizing. Default is
NULL, ie no grouping variable is used. Note that if no grouping variable is
used, the returned data.table will still contain a ".by" variable, taking
the value given in \code{by_total_val}}

\item{by_total_val}{String. Name given to overall by-group.}
}
\value{
A data.table containing a descriptive summary of \code{var}. The
data.table contains the following variables:
Group information:
\itemize{
\item ".var_name": Variable name.
\item ".var_type": Variable type.
\item ".by": by-group value.
Summary statistics:
\item ".n_by_level": Number of observations in ".by" group.
\item ".var_level": Variable level. Only used for categorical variables.
\item ".n_var_level": Number of observations in ".var_level" group.
\item ".sum": Sum of variable in group. For binary and categorical variables
.n_var_level = .sum.
\item .stddev: Standard deviation in group. NA for binary and categorical
variables.
\item ".p25": 25th percentile.
\item ".p50": 50th percentile / median.
\item ".p75": 75th percentile.
}
}
\description{
Creates a descriptive summary of a variable in a data.frame.
}
\details{
If \code{type} is NULL, the variable type of \code{var} will be deduced from the data:
\enumerate{
\item If the variable takes "binary" values, ie is a logical vector, or a numeric
vector with zero and one's, then \code{type} is set to "bin".
\item Else, if the variable is numeric then \code{type} is set to "cont".
\item Else, \code{type} is set to "cat".
}
}
\examples{
\dontrun{
  x <- data.frame(var = 1:10)
  .summarize_var(x, "var")
}
}
\keyword{internal}
