% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{.summarize_var}
\alias{.summarize_var}
\title{Summarize variable}
\usage{
.summarize_var(
  x,
  var,
  type = NULL,
  weight = NULL,
  by = NULL,
  by_total_val = ".all"
)
}
\arguments{
\item{x}{Input data.frame.}

\item{var}{String. Name of variable in \code{x} to summarize.}

\item{type}{String. Variable type of \code{var}: "bin" (binary),
"cont" (continuous), "cat" (categorical). Default is NULL, meaning that
the type of \code{var} is automatically deduced from the data.}

\item{weight}{String. Optional name of variable in \code{x} that contains
observation weights. The weights must be non-negative numbers.}

\item{by}{String. Optional name of variable to summarize by.}

\item{by_total_val}{String. Name given to overall by-group.}
}
\value{
A data.table containing a descriptive summary of \code{var}.
}
\description{
Returns a descriptive summary of a variable in a data.frame.
}
\examples{
\dontrun{
  x <- data.frame(var = 1:10)
  .summarize_var(x, "var")
}
}
\keyword{internal}
