% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mask.R
\name{.update_mask_flags}
\alias{.update_mask_flags}
\title{title}
\usage{
.update_mask_flags(
  x,
  mask_flags = rep(FALSE, length(x)),
  protect = rep(FALSE, length(x)),
  mask_min = 1L,
  mask_max = 4L,
  mask_avg = 1L,
  max_ite = 100L
)
}
\arguments{
\item{x}{a}

\item{mask_flags}{a}

\item{mask_min}{a}

\item{mask_max}{a}

\item{mask_avg}{a}
}
\value{
a
}
\description{
TODO:
\itemize{
\item NA's in x should be considered invaild input? There is no case where that
makes sense? and that assumption is also used in code.
\item Loop until nothing more is flagged for masking.
\item Change function so that input is a vector and a corresponding
vector with flags for which elements are do be flagged (so far).
the function then updates the flag vector and returns that updated
vector of flags
\item Function should also be able to take a vector with flags for values that
are protected from masking, eg zeros that are known to be zeros and therefore are not valid to mask
(very relevant, something we want to be able to control)
\item Protect ruins things, if protected cell NEEDS to be masked, eg if count under 5?
function will loop until max iteration reached? Abandon approach? If the protected
cells leads to problems then function can return warning/error?
}
}
\examples{
\dontrun{TODO}
}
\keyword{internal}
