% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wmean}
\alias{.wmean}
\title{Weighted arithmetic mean}
\usage{
.wmean(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}. Default is NULL, ie no weights are used.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
A numeric value.
}
\description{
Calculates the weighted arithmetic mean of a numeric/logical vector \code{x} with
associated weights \code{w}.
}
\examples{
\dontrun{
  .wmean(x = 1:2, w = 1:2)
}
}
\keyword{internal}
