% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wmean}
\alias{.wmean}
\title{Weighted arithmetic mean}
\usage{
.wmean(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with \code{x}.}

\item{na.rm}{Boolean. Should NA values in \code{x} and values in \code{x} with
NA weights in \code{w} be removed?}
}
\value{
A numeric.
}
\description{
Calculates the weighted arithmetic mean of a vector \code{x} with
associated weights \code{w}.
}
\examples{
\dontrun{
   x <- 1:5
   w <- 1:5
  .wmean(x, w)
}
}
\keyword{internal}
