% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wquantile}
\alias{.wquantile}
\title{Weighted quantile}
\usage{
.wquantile(x, w = rep(1L, length(x)), prob, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}.}

\item{prob}{Numeric vector of probabilities with values in \verb{[0;1]}
corresponding to quantiles to estimate.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
A numeric vector.
}
\description{
Estimates quantiles using the weighted Harrell-Davis quantile estimator.
}
\details{
Implemented based on https://arxiv.org/pdf/2304.07265.pdf.
}
\examples{
\dontrun{
  x <- 1:5
  w <- c(1, 0, 1, 0, 1)
  .wquantile(x, w, prob = 0.5)
}
}
\keyword{:}
\keyword{internal}
