% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wsd}
\alias{.wsd}
\title{Weighted standard deviation}
\usage{
.wsd(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with \code{x}.}

\item{na.rm}{Boolean. Should NA values in \code{x} and values in \code{x} with
NA weights in \code{w} be removed?}
}
\value{
A numeric.
}
\description{
Calculates the weighted standard deviation(sd) of a vector \code{x} with
associated weights \code{w}.
}
\examples{
\dontrun{
   x <- 1:5
   w <- rep(2, 5)
  .wsd(x, w)
}
}
\keyword{internal}
