% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wsd}
\alias{.wsd}
\title{Weighted standard deviation}
\usage{
.wsd(x, w = rep(1L, length(x)), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
A numeric value.
}
\description{
Calculates the weighted standard deviation(sd) of a vector \code{x} with
associated weights \code{w}.
}
\details{
The standard deviation of a length-one or zero-length vector is NA.
}
\examples{
\dontrun{
  .wsd(x = 1:5)
}
}
\keyword{internal}
