% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wsum}
\alias{.wsum}
\title{Weighted sum of vector elements}
\usage{
.wsum(x, w = rep(1L, length(x)), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical vector.}

\item{w}{A numeric vector of non-negative numbers. Weights associated with
\code{x}.}

\item{na.rm}{A logical value. Should missing values (including NaN) in \code{x}
be removed?}
}
\value{
A numeric.
}
\description{
Calculates the weighted sum of a numeric vector \code{x} with associated
weights \code{w}.
}
\examples{
\dontrun{
   .wsum(x = 1:2, w = rep(2, 2))
}
}
\keyword{internal}
