% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{summarize_df}
\alias{summarize_df}
\title{Summarizes variables in a data.frame}
\usage{
summarize_df(x, vars, weight = NULL, by = NULL, by_total_val = ".all")
}
\arguments{
\item{x}{Input data.frame.}

\item{vars}{Character vector. Names of variables in \code{x} to summarize.
Variables types can be specified by naming elements of \code{var}. Variable
types are: "bin" (binary), "cont" (continuous), "cat" (categorical).
If ny variable type is specified, the type is automatically deduced from
the data.}

\item{weight}{String. Name of variable in \code{x} that contains
observation weights. Must be a numeric variable, with non-negative
numbers. Default is NULL, ie no weights are used.}

\item{by}{String. Name of variable to group by while summarizing. Default is
NULL, ie no grouping variable is used.}

\item{by_total_val}{String. Name given to overall by-group.}
}
\value{
A data.table containing a descriptive summary of \code{vars}.
}
\description{
Returns a descriptive summary of variables in a data.frame.
}
\examples{
\dontrun{
  x <- data.frame(var1 = 1:10, var2 = c(rep("A", 5), rep("B", 5)))
  summarize_df(
   x,
   vars = c(cont = "var1", "var2")
  )
}
}
\keyword{internal}
