% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{Summarize variables}
\usage{
tbl_summary(
  x,
  vars,
  weight = NULL,
  by = NULL,
  by_total_val = ".all",
  mask = FALSE,
  add_num_stat = FALSE
)
}
\arguments{
\item{x}{a}

\item{vars}{a}

\item{weight}{a}

\item{by}{a}

\item{by_total_val}{a}

\item{mask}{A logical value. Mask person-sensitive aggregated counts?}

\item{add_num_stat}{A logical value. Add numeric columns with statistics?}
}
\value{
data.table
}
\description{
Summarize variables
}
\examples{
\dontrun{
  TODO
}
}
