% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{Summarize variables}
\usage{
tbl_summary(
  x,
  vars,
  types = NULL,
  weight = NULL,
  by = NULL,
  by_total_val = ".all"
)
}
\arguments{
\item{x}{a}

\item{vars}{a}

\item{types}{a}

\item{weight}{a}

\item{by}{a}

\item{by_total_val}{a}
}
\value{
data.table
}
\description{
Summarize variables
}
\examples{
\dontrun{
  TODO
}
}
